+++ packages/eyes-api/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- renamed TestResultsContainer to TestResultContainer
 
 ## 0.0.3 - 2021/4/12